Add redirect_url_for (optional) to is_active_feature
It's not possible to use url_for() in the decorator call if the app context is
not build yet. Therefore this patch adds a new optional argument named
redirect_url_for to is_active_feature. It's the same as redirect_to but it
triggers url_for() right before the redirect.